SL_Project/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Marshmallow 
marshmallow-sqlalchemy
beautifulsoup4 
selenium 
python-dateutil
